backend/pyproject.toml
[project]
name = "contentforge"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["pytest-asyncio"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
